requirements.txt
pandas==2.3.3
pinecone==7.3.0
dotenv==0.9.9
fastapi==0.121.1
tiktoken==0.12.0
pedantic==2.12.4
sqlite-utils==3.38
langchain-openai==1.0.2
torch==2.9.1+cpu
sentence-transformers==5.1.2
streamlit==1.51.0